Fix keys errors
I had to solve the problem. It was starting to annoy me! Sorry!

I moved the component to be drumpads and moved the map inside as it kept
trying to render each of the pads all the time. This is purely for
performance.

I removed some of the unnecessary props and tidied it up a bit.

The important bit is the keypress.

The useEffect is called only once and the reference to the window event
is available to be called whenever then.

I passed the function for handling the keypress in as a reference.

The es-lint comment is to stop it complaining about the need for it to
go into the dependency array (the only reason i chose this rather than
to add it as it then wants the other functions in there too and that
starts to get all sorts of crazy). There's probably other ways to solve
that nicer too, but I ultimately wanted to solve the keypress and solve
the issue with things overlapping with async/await so you could see how
that worked.

I extracted the reset functionality out too so you can see how they're
chained in a way that allows them to happen as you'd expect. Stop if
playing, reset that time and then play the next (whether that's the same
or not).

Hope this is useful. It was bothering me too much haha!